// src/typings/productCategories.d.ts
export type ProductCategory =
  | 'Vegetables'
  | 'Fruits'
  | 'Dairy'
  | 'Meats'
  | 'Fishes'
  | 'Seafoods'
  | 'DairyAlternatives'
  | 'DairyAndMeatAlternatives'
  | 'BakingAndGrains'
  | 'Legumes'
  | 'Nuts'
  | 'Spices'
  | 'Oils'
  | 'Sauces'
  | 'Beverages'
  | 'Alcohol'
  | 'DessertsAndSnacks'
  | 'FoodAdditives'
  | 'Sweeteners'
  | 'Others'